JIRA:SMUPD-59 Some minor changes were made to get rid of compiler warnings
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@82957 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/wgs_parse/wgs_utils.cpp
@@ -55,8 +55,8 @@ CScope& GetScope()
     return scope;
 }
 
-static const size_t MAX_SEVEN_DIGITS_NUM = 9999999;
-static const size_t MAX_SIX_DIGITS_NUM = 999999;
+static const int MAX_SEVEN_DIGITS_NUM = 9999999;
+static const int MAX_SIX_DIGITS_NUM = 999999;
 
 size_t GetMaxAccessionLen(int accession_num)
 {
